Change linting strategy for non-model factories
We had been detecting non-model factories (hashes and example
SearchWorks items) based on the name of the factory so that
we could avoid trying to lint them with FactoryBot. This changes
the strategy to just directly check if the factory's built class
is of a certain type, which is more predictable.