Use paging as a fallback if any items are requestable
This handles the case where we can't conclusively determine which
request type to delegate to, because each of the items would be
requested a different way. As a fallback, we assume it's a page
and render the item selector so that the items can figure it out
later depending on which one we ask for.