Request the folio instance hrid so it can be used in links for Aeon
Fixes #1890

We use the hrid to make the "view in searchworks" link that we send
to Aeon when creating a page request, but we weren't actually
requesting it from graphql, so the links didn't go anywhere useful.

Also adds a test for this property to confirm we correctly strip
the leading 'a' from hrids when present.