Prevent the item selector search from submitting the form on enter
This solution creates a hidden, non-interactable submit button
at the beginning of the form, which becomes its "default submit"
button according to the HTML spec. Because hitting enter will
attempt to activate this button and it's disabled, nothing happens.

If you have the final, real submit button focused and hit enter,
that will work, because the explicit target of that action is the
(non-disabled) submit button.

This ensures that no elements in the form are allowed to submit
it accidentally except the final submit button.

Fixes #2364